Message.TimeToLive should be an int #8

+++ Code/Message.cs
@@ -59,7 +59,7 @@ namespace FCM.Net
         /// This parameter specifies how long (in seconds) the message should be kept in FCM storage if the device is offline. The maximum time to live supported is 4 weeks, and the default value is 4 weeks. 
         /// </summary>
         [JsonProperty("time_to_live")]
-        public byte? TimeToLive { get; set; }
+        public int? TimeToLive { get; set; }
 
         /// <summary>
         /// This parameter specifies the package name of the application where the registration tokens must match in order to receive the message.
